msm: apuclock-krait: Introduce simple VDD interface
Use the newly added CPUFreq API to create a voltage-on-frequency table
accessible to users.  This can help with a stress balancing to improve
cpu stability, especially on edge frequencies, which are used literally
always.

Krait power subsystem uses vdd_core variable to set a cpu core voltage.
Now this value can be changed manually from userspace. acpuclk-krait
returns the VDD table in the style of "frequency: voltage":
 ...
 1497mhz: 930 mV
 1267mhz: 895 mV
 1190mhz: 885 mV
 ...

Voltage values are writable and can be changed by passing the whole
table to UV_mV_table sysfs node. That's not convenient for a terminal
editing, but is evaluated by third-party applications and makes
the modification via text editor easier.

Both sysfs funtions report only the actual frequencies used, leaving
those that are disabled in acpuclk table abroad.  In addition, write
function ensures that the passed value is within the limits.  Note that
the lower limit is a simple zero.  Although Krait regulator provides an
accurate minimal voltage limit, other regulators lack this info. That's
why I've decided to leave it on a user's conscience.

Change-Id: I79c279523268953b72139e9480e00b31ddbf1a36
Signed-off-by: Alex Saiko <solcmdr@gmail.com>